pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.3.3", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "tvm-ffi-object-loader"
version = "0.1.0"
description = "Load TVM-FFI exported object files using LLVM ORC JIT v2"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "TVM-FFI Object Loader Contributors"}
]
keywords = ["tvm-ffi", "llvm", "jit", "object-loader"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C++",
]
dependencies = [
    "apache-tvm-ffi>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yaoyaoding/demo-object-loading"
Repository = "https://github.com/yaoyaoding/demo-object-loading"

[tool.scikit-build]
cmake.minimum-version = "3.18"
cmake.build-type = "Release"
wheel.py-api = "py3"
# Don't set install-dir, let it use the default python/ directory

[tool.scikit-build.cmake.define]
CMAKE_EXPORT_COMPILE_COMMANDS = "ON"
